moved reusable stuff out to utils
