+++ src/frameworks/react/reactTailwindApp.js
@@ -1,7 +1,11 @@
 const shell = require('shelljs');
-const fse = require('fs-extra');
 const chalk = require('chalk');
 const config = require('./reactTailwindConfig');
+const {
+  installDependencies,
+  updatePackageScripts,
+  addTemplates,
+} = require('../../utils');
 
 const createReactApp = (name) => {
   return new Promise((resolve, reject) => {
@@ -17,57 +21,7 @@ const createReactApp = (name) => {
   });
 };
 
-const installDependencies = async (dependencies, devDependencies) => {
-  await new Promise((resolve) => {
-    console.log('Installing tailwind dependencies...');
-    shell.exec(`yarn add ${dependencies.join(' ')}`, () => {
-      resolve();
-    });
-  });
-
-  await new Promise((resolve) => {
-    console.log('Installing tailwind dev dependencies...');
-    shell.exec(`yarn add --dev ${devDependencies.join(' ')}`, () => {
-      resolve();
-    });
-  });
-};
-
-const updatePackageScripts = async (scripts) => {
-  console.log('Updating package scripts...');
-  return new Promise((resolve) => {
-    const basePackage = JSON.parse(fse.readFileSync('package.json'));
-
-    scripts.forEach((script) => {
-      basePackage.scripts[script.key] = script.value;
-    });
-
-    fse.writeFile(
-      'package.json',
-      JSON.stringify(basePackage, null, 2),
-      function (err) {
-        if (err) {
-          return console.log(err);
-        }
-
-        console.log('Scripts updated');
-        resolve();
-      },
-    );
-  });
-};
-
-const addTemplates = (templates) => {
-  return new Promise((resolve) => {
-    templates.forEach((template) => {
-      fse.outputFileSync(template.path, template.file);
-    });
-
-    resolve();
-  });
-};
-
-exports.create = async (name, directory) => {
+exports.create = async (name) => {
   await createReactApp(name);
   await installDependencies(config.dependencies, config.devDependencies);
   await updatePackageScripts(config.scripts);
